backend/requirements.txt
charset-normalizer==3.4.4
pillow==12.1.0
reportlab==4.4.9
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.1  # For PostGIS support

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Real-time & Background
python-socketio==5.10.0
aiohttp==3.9.1
celery==5.3.4
redis==5.0.1

stripe>=6.0.0,<7.0.0  
fastapi-mail>=0.4.0
aiosmtplib>=2.0.0
reportlab==4.4.9

# Geospatial & Maps
geopy==2.4.0  # For distance calculations
shapely==2.0.2  # For geometry operations

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Background Tasks & PDF
weasyprint==61.0  # For PDF receipts
celery-redbeat==0.13.0  # For scheduled tasks

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

python-depends

fastapi-mail
email-validator
psutil
